move whoami command
