+++ docs/cmd/www/meroxa-add-resource.md
@@ -1,6 +1,6 @@
 ---
-createdAt: 2021-04-16T17:48:56+02:00
-updatedAt: 2021-04-16T17:48:56+02:00
+createdAt: 2021-04-16T19:26:31+02:00
+updatedAt: 2021-04-16T19:26:31+02:00
 title: "meroxa add resource"
 slug: meroxa-add-resource
 url: /cli/meroxa-add-resource/
@@ -14,7 +14,7 @@ Add a resource to your Meroxa resource catalog
 Use the add command to add resources to your Meroxa resource catalog.
 
 ```
-meroxa add resource [NAME] --type TYPE [flags]
+meroxa add resource [NAME] --type TYPE --url URL [flags]
 ```
 
 ### Examples
@@ -36,7 +36,7 @@ meroxa add resource slack --type url -u $WEBHOOK_URL
       --client-key string    client private key for authenticating to the resource
   -h, --help                 help for resource
   -m, --metadata string      resource metadata
-      --password string      passsword
+      --password string      password
       --ssl                  use SSL
       --type string          resource type
   -u, --url string           resource url
