+++ docs/cmd/www/meroxa-logs-connector.md
@@ -1,6 +1,6 @@
 ---
-createdAt: 2021-04-16T17:48:56+02:00
-updatedAt: 2021-04-16T17:48:56+02:00
+createdAt: 2021-04-16T19:26:31+02:00
+updatedAt: 2021-04-16T19:26:31+02:00
 title: "meroxa logs connector"
 slug: meroxa-logs-connector
 url: /cli/meroxa-logs-connector/
