+++ docs/cmd/www/meroxa-remove-endpoint.md
@@ -1,6 +1,6 @@
 ---
-createdAt: 2021-04-16T17:48:56+02:00
-updatedAt: 2021-04-16T17:48:56+02:00
+createdAt: 2021-04-16T19:26:31+02:00
+updatedAt: 2021-04-16T19:26:31+02:00
 title: "meroxa remove endpoint"
 slug: meroxa-remove-endpoint
 url: /cli/meroxa-remove-endpoint/
